Filled in some of the functions for segment stitching
IR sensors now defined as objects for ease of use (maybe)

Conversion of segments to measurements and individual SegmentPoints to
measurements is being done, but not sure if it is working correctly yet. Seems
to generate something like what is to be expected.

Need to use a frame to visualise the pointcloud that is being temporarily
published as a debug tool. Need to remember to define a map frame.